Split circulars server-side code off into a separate file
This reduces the size of the client code bundle.

Before:

    $ du -hc public/build
    4.3M    public/build/_assets
    3.3M    public/build/_shared
     24K    public/build/routes/circulars
     20K    public/build/routes/__auth
    128K    public/build/routes/missions
     12K    public/build/routes/user/email
    8.0K    public/build/routes/user/credentials
     48K    public/build/routes/user
    148K    public/build/routes/docs
     16K    public/build/routes/quickstart
    452K    public/build/routes
    8.1M    public/build
    8.1M    total

Afer:

    lpsinger$ du -hc public/build/
    4.3M    public/build//_assets
    1.6M    public/build//_shared
     24K    public/build//routes/circulars
     20K    public/build//routes/__auth
    128K    public/build//routes/missions
     12K    public/build//routes/user/email
    8.0K    public/build//routes/user/credentials
     48K    public/build//routes/user
    148K    public/build//routes/docs
     16K    public/build//routes/quickstart
    452K    public/build//routes
    6.4M    public/build/
    6.4M    total
